requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-dotenv
python-jose
python-multipart
